refactor(#3038): move tmux watcher commit decisions
